# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Jewish prayer times calculation and publisher management",
        "title": "Zmanim Lab API",
        "termsOfService": "https://zmanim-lab.com/terms",
        "contact": {
            "name": "Zmanim Lab Support",
            "email": "support@zmanim-lab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/publisher/cache": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Clears all cached zmanim calculations for the authenticated publisher",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publisher"
                ],
                "summary": "Invalidate publisher cache",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Publisher ID",
                        "name": "X-Publisher-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cache invalidated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Publisher not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/zmanim": {
            "get": {
                "description": "Calculates Jewish prayer times (zmanim) for a specific city and date, optionally using a publisher's custom algorithm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zmanim"
                ],
                "summary": "Get zmanim for a city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City ID from the cities database",
                        "name": "cityId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Publisher ID for custom algorithm (uses default if not specified)",
                        "name": "publisherId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format (defaults to today)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculated zmanim with formula details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handlers.ZmanimWithFormulaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Calculates zmanim using raw latitude/longitude coordinates. Prefer the GET /zmanim endpoint with cityId for better accuracy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zmanim"
                ],
                "summary": "Calculate zmanim by coordinates (legacy)",
                "parameters": [
                    {
                        "description": "Coordinates and date",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jcom-dev_zmanim-lab_internal_models.ZmanimRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculated zmanim",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handlers.ZmanimWithFormulaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_jcom-dev_zmanim-lab_internal_models.ZmanimRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "YYYY-MM-DD format",
                    "type": "string"
                },
                "elevation": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "publisher_id": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_handlers.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/internal_handlers.ResponseMeta"
                }
            }
        },
...
        "internal_handlers.FormulaDetails": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "explanation": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "internal_handlers.ResponseMeta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.ZmanWithFormula": {
            "type": "object",
            "properties": {
                "formula": {
                    "$ref": "#/definitions/internal_handlers.FormulaDetails"
                },
                "is_beta": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "time_category": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.ZmanimLocationInfo": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "string"
                },
                "city_name": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "region": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.ZmanimPublisherInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.ZmanimWithFormulaResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean"
                },
                "cached_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/internal_handlers.ZmanimLocationInfo"
                },
                "publisher": {
                    "$ref": "#/definitions/internal_handlers.ZmanimPublisherInfo"
                },
                "zmanim": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handlers.ZmanWithFormula"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token from Clerk authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}